refactor: move threading and mutex to own module
