+++ helpers/utility.py
@@ -12,15 +12,11 @@ from config.settings import WEBROB_PATH
 from flask_user import current_user
 from flask_user import current_app
 from functools import wraps
-from threading import Lock, Thread
 from pathlib2 import Path
-from concurrent.futures import ThreadPoolExecutor
 
 from app_and_db import app
 from helpers.file_handler import move_file, write_binary_file
 from Crypto.Random import random
-
-THREAD_POOL_EXECUTOR = ThreadPoolExecutor(max_workers=3)
 
 def get_user_dir():
     userDir = "/home/ros/user_data/" + session['user_container_name']
@@ -70,42 +66,6 @@ def download_file(url, file_path):
                 Path(temp_downloads_dir).rmdir()
         except Exception as e:
             app.logger.info('Could not remove temp dir.\n\n' + e.__str__())
-
-
-def start_thread(target_func):
-    THREAD_POOL_EXECUTOR.submit(target_func)
-
-
-def mutex_lock(mutex=None):
-    """ decorator function that provides a mutex-lock
-    
-    If no mutex lock is assigned, a per-function mutex is used.
-    
-    The lock is thread-blocking, so in order to not block
-    the current thread create a new one to run the desired
-    function. You can do that, for example, with "start_thread" """
-    
-    def wrap_inner(func):
-        p_mutex = _determine_mutex(mutex, func)
-
-        @wraps(func)
-        def inner(*args, **kwargs):
-            p_mutex.acquire()
-            try:
-                func(*args, **kwargs)
-            finally:
-                p_mutex.release()
-
-        return inner
-    return wrap_inner
-
-
-def _determine_mutex(mutex, func):
-    if mutex is None:
-        func.lock = Lock()
-        return func.lock
-    else:
-        return mutex
 
 
 def type_str(obj):
