+++ pages/neem_overview.py
@@ -11,8 +11,9 @@ from html_sanitizer import Sanitizer
 from html_sanitizer.sanitizer import sanitize_href, bold_span_to_strong,italic_span_to_em, target_blank_noopener, tag_replacer
 
 from config.settings import WEBROB_PATH, STATIC_DIR_PATH, DEFAULT_FILES_PATH, DOWNLOADS_DIR_PATH
-from helpers.utility import download_file, mutex_lock, start_thread
+from helpers.utility import download_file
 from helpers.file_handler import copy_file, copy_dir, remove_if_is_dir, remove_if_is_file, unzip_file, dump_dict_to_json, get_dict_from_json, read_file, write_non_binary_file, make_archive_of_files_and_dirs
+from helpers.threading import start_thread, mutex_lock
 from neems.neemhub import instance as neemhub, NEEMHubConnectionError
 from neems.neem import DEFAULT_IMAGE_PATH, DEFAULT_IMAGE_PATH_NO_STATIC
 
@@ -125,7 +126,7 @@ def automatic_update_neem_overview_files():
     ContentSettings.set_last_update_type_neem_overview(UpdateMethod.AUTOMATIC)
 
 
-# ideally only start this function in a seperate thread with utility.start_thread
+# ideally only start this function in a seperate thread with threading.start_thread
 # exception is the app start-up
 @mutex_lock(OVERVIEW_MUTEX)
 def _update_neem_overview_files():
